[GATE-80] Fixed oracle schema. Updated for 1.0 Beta 3a release.
git-svn-id: http://svn.igniterealtime.org/svn/repos/wildfire/trunk/src/plugins/gateway@5590 b35dd754-fafc-0310-a699-88a17e54d16e

+++ src/database/gateway_oracle.sql
@@ -13,7 +13,7 @@ CREATE INDEX gatewayReg_jid_idx ON gatewayRegistration (jid);
 CREATE INDEX gatewayReg_type_idx ON gatewayRegistration (transportType);
 
 CREATE TABLE gatewayPseudoRoster (
-   registrationID    BIGINT         NOT NULL,
+   registrationID    INTEGER        NOT NULL,
    username          VARCHAR2(255)  NOT NULL,
    nickname          VARCHAR2(255),
    groups            VARCHAR2(255)
